Added EmployerRegistration and styling
I'm leaving for work in a couple but I'm gonna keep working on this
tonight. I'm gonna finish giving users the ability to sign up and then
I'll work on managing users on the rest of the site. If anyone wants to
help let me know, there's gonna be alot to it.